syosetu/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
cloudscraper>=1.2.71
python-dotenv>=1.0.0
